refactor: created global component for kanban board
