--- apps/app/components/common/board-view/single-board.tsx
@@ -1,3 +1,0 @@
-const SingleBoard = () => <></>;
-
-export default SingleBoard;
