+++ apps/app/components/issues/board-view.tsx
@@ -5,16 +5,14 @@ import { useRouter } from "next/router";
 import useSWR, { mutate } from "swr";
 
 // react-beautiful-dnd
-import { DragDropContext, DropResult } from "react-beautiful-dnd";
+import { DragDropContext, Draggable, DropResult } from "react-beautiful-dnd";
 // hook
-import useIssuesProperties from "hooks/use-issue-properties";
 import useIssueView from "hooks/use-issue-view";
 // services
 import stateServices from "services/state.service";
 import issuesServices from "services/issues.service";
-import projectService from "services/project.service";
 // components
-import SingleBoard from "components/issues/board-view/single-board";
+import { CommonSingleBoard } from "components/core/board-view/single-board";
 import StrictModeDroppable from "components/dnd/StrictModeDroppable";
 import { CreateUpdateIssueModal, DeleteIssueModal } from "components/issues";
 // ui
@@ -22,7 +20,7 @@ import { Spinner } from "components/ui";
 // types
 import type { IState, IIssue, IssueResponse, UserAuth } from "types";
 // fetch-keys
-import { STATE_LIST, PROJECT_ISSUES_LIST, PROJECT_MEMBERS } from "constants/fetch-keys";
+import { STATE_LIST, PROJECT_ISSUES_LIST } from "constants/fetch-keys";
 
 type Props = {
   issues: IIssue[];
@@ -30,7 +28,7 @@ type Props = {
   userAuth: UserAuth;
 };
 
-const BoardView: React.FC<Props> = ({ issues, handleDeleteIssue, userAuth }) => {
+export const IssuesBoardView: React.FC<Props> = ({ issues, handleDeleteIssue, userAuth }) => {
   const [createIssueModal, setCreateIssueModal] = useState(false);
   const [isIssueDeletionOpen, setIsIssueDeletionOpen] = useState(false);
   const [issueDeletionData, setIssueDeletionData] = useState<IIssue | undefined>();
@@ -49,21 +47,6 @@ const BoardView: React.FC<Props> = ({ issues, handleDeleteIssue, userAuth }) => 
       ? () => stateServices.getStates(workspaceSlug as string, projectId as string)
       : null
   );
-
-  const { data: members } = useSWR(
-    workspaceSlug && projectId ? PROJECT_MEMBERS(projectId as string) : null,
-    workspaceSlug && projectId
-      ? () => projectService.projectMembers(workspaceSlug as string, projectId as string)
-      : null,
-    {
-      onErrorRetry(err, _, __, revalidate, revalidateOpts) {
-        if (err?.status === 403) return;
-        setTimeout(() => revalidate(revalidateOpts), 5000);
-      },
-    }
-  );
-
-  const [properties] = useIssuesProperties(workspaceSlug as string, projectId as string);
 
   const handleOnDragEnd = useCallback(
     (result: DropResult) => {
@@ -208,36 +191,42 @@ const BoardView: React.FC<Props> = ({ issues, handleDeleteIssue, userAuth }) => 
                     ref={provided.innerRef}
                   >
                     <div className="flex h-full gap-x-4 overflow-x-auto overflow-y-hidden">
-                      {Object.keys(groupedByIssues).map((singleGroup, index) => (
-                        <SingleBoard
-                          key={singleGroup}
-                          selectedGroup={selectedGroup}
-                          groupTitle={singleGroup}
-                          createdBy={
-                            selectedGroup === "created_by"
-                              ? members?.find((m) => m.member.id === singleGroup)?.member
-                                  .first_name ?? "loading..."
-                              : null
-                          }
-                          groupedByIssues={groupedByIssues}
-                          index={index}
-                          setIsIssueOpen={setCreateIssueModal}
-                          properties={properties}
-                          setPreloadedData={setPreloadedData}
-                          stateId={
-                            selectedGroup === "state_detail.name"
-                              ? states?.find((s) => s.name === singleGroup)?.id ?? null
-                              : null
-                          }
-                          bgColor={
-                            selectedGroup === "state_detail.name"
-                              ? states?.find((s) => s.name === singleGroup)?.color
-                              : "#000000"
-                          }
-                          handleDeleteIssue={handleDeleteIssue}
-                          userAuth={userAuth}
-                        />
-                      ))}
+                      {Object.keys(groupedByIssues).map((singleGroup, index) => {
+                        const stateId =
+                          selectedGroup === "state_detail.name"
+                            ? states?.find((s) => s.name === singleGroup)?.id ?? null
+                            : null;
+
+                        return (
+                          <Draggable key={singleGroup} draggableId={singleGroup} index={index}>
+                            {(provided, snapshot) => (
+                              <CommonSingleBoard
+                                provided={provided}
+                                snapshot={snapshot}
+                                bgColor={
+                                  selectedGroup === "state_detail.name"
+                                    ? states?.find((s) => s.name === singleGroup)?.color
+                                    : "#000000"
+                                }
+                                groupTitle={singleGroup}
+                                groupedByIssues={groupedByIssues}
+                                selectedGroup={selectedGroup}
+                                addIssueToState={() => {
+                                  setCreateIssueModal(true);
+                                  if (selectedGroup)
+                                    setPreloadedData({
+                                      state: stateId !== null ? stateId : undefined,
+                                      [selectedGroup]: singleGroup,
+                                      actionType: "createIssue",
+                                    });
+                                }}
+                                handleDeleteIssue={handleDeleteIssue}
+                                userAuth={userAuth}
+                              />
+                            )}
+                          </Draggable>
+                        );
+                      })}
                     </div>
                     {provided.placeholder}
                   </div>
@@ -254,5 +243,3 @@ const BoardView: React.FC<Props> = ({ issues, handleDeleteIssue, userAuth }) => 
     </>
   );
 };
-
-export default BoardView;
