+++ apps/app/components/issues/index.ts
@@ -1,7 +1,7 @@
-export * from "./board-view";
 export * from "./comment";
 export * from "./sidebar-select";
 export * from "./activity";
+export * from "./board-view";
 export * from "./delete-issue-modal";
 export * from "./description-form";
 export * from "./form";
